pyproject.toml
[project]
name = "forgeai"
version = "0.2.0"
description = "ForgeAI - Multi-Agent Adaptive Research System with Hierarchical Context Engineering"
authors = [{ name = "Orimoloye", email = "your.email@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Core LangChain & LangGraph
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.0",           # Modern Chroma integration
    
    # LLM Providers
    "langchain-groq>=0.2.0",
    "langchain-google-genai>=4.0.0",
    
    # Rerankers
    "voyageai>=0.2.0",
    "cohere>=5.0.0",
    
    # Embeddings & Vector Store
    "langchain-huggingface>=0.1.0",
    "chromadb>=0.5.0",
    
    # Document Loading
    "pypdf>=5.0.0",
    
    # Search & Tools
    "tavily-python>=0.5.0",
    "arxiv>=2.1.0",
    
    # Utilities
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    
    # UI & Dashboard
    "streamlit>=1.38.0",
    "plotly>=5.24.0",
    "pandas>=2.2.0",
    
    # Evaluation
    "ragas>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.8.0",
    "isort>=5.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "tqdm>=4.66.0",
]

[tool.poetry]  # Optional: if you prefer using Poetry

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
multi_line_output = 3